Move cfpropertylist to fastlane_core/core_ext. Use in ipa_file_analyzer to avoid system call.
