+++ fastlane_core/lib/fastlane_core/core_ext/cfpropertylist.rb
